backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Nathan Yeager <nyeager16@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
django = "^5.2"
djangorestframework = "^3.16.0"
psycopg = {extras = ["binary"], version = "^3.2.3"}
dotenv = "^0.9.9"
pandas = "^2.2.3"
stanza = "^1.10.1"
scrapetube = "^2.5.1"
youtube-transcript-api = "^1.0.3"
django-extensions = "^4.1"
pydot = "^3.0.4"
fsrs = "^5.1.3"
django-cors-headers = "^4.7.0"
djangorestframework-simplejwt = "^5.5.0"
django-background-tasks = "^1.2.8"
deep-translator = "^1.11.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
